+++ src/app/(route)/(blog)/blog/[username]/_components/PostContainer.tsx
@@ -2,18 +2,13 @@
 
 import { useEffect, useState } from 'react';
 import { useParams, useRouter } from 'next/navigation';
-import { useQueryClient } from '@tanstack/react-query';
 import { v4 as uuidv4 } from 'uuid';
 import Post from '../../../../../_components/Post';
-import { fetchPost, fetchProfile, putBookmark, putLike } from '../_services/blogService';
+import { fetchPost, fetchProfile } from '../_services/blogService';
 import useCustomQuery from '@/app/_hooks/useCustomQuery';
 import { PostDetailProps } from '@/app/(route)/(main-page)/main/_types/main-page';
 import { formatDate } from '@/app/_utils/formatDate';
-import useCustomMutation from '@/app/_hooks/useCustomMutation';
-
-type MutationContext = {
-  previousPosts: PostDetailProps[];
-};
+import useLoveAndBookmark from '@/app/_hooks/useLoveAndBookmark';
 
 export default function PostContainer() {
   const router = useRouter();
@@ -23,7 +18,6 @@ export default function PostContainer() {
 
   const memberId = memberData?.memberId;
   const [lastId, setLastId] = useState('0');
-  const queryClient = useQueryClient();
 
   const { data, isLoading } = useCustomQuery(['post', memberId], () => fetchPost(memberId, lastId));
 
@@ -35,68 +29,7 @@ export default function PostContainer() {
     }
   }, [data]);
 
-  const loveMutation = useCustomMutation<unknown, number>(putLike, {
-    // mutation 실행 전에 실행되는 onMutate
-    onMutate: async (postId) => {
-      // 진행중인 post 쿼리를 취소하여 충돌 방지
-      await queryClient.cancelQueries({ queryKey: ['post', memberId] });
-      const previousPosts = [...posts]; // 롤백을 위한 현재 posts 상태 저장
-      // 낙관적 업데이트, UI 즉시 업데이트
-      setPosts((prevPosts) =>
-        prevPosts.map((post) => {
-          if (post.postId === postId) {
-            return {
-              ...post,
-              isLoved: !post.isLoved, // 좋아요 상태 토글
-              lovedCount: post.isLoved ? post.lovedCount - 1 : post.lovedCount + 1,
-            };
-          }
-          return post;
-        }),
-      );
-
-      return { previousPosts } as MutationContext; // 이전 상태를 context로 반환하여 오류 발생 시 복구
-    },
-    onError: (_, __, context) => {
-      if ((context as MutationContext)?.previousPosts) {
-        setPosts((context as MutationContext).previousPosts);
-      }
-    },
-    // mutationd 완료 후 실행되는 함수
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['post', memberId] });
-    },
-  });
-
-  const bookmarkMutation = useCustomMutation<unknown, number>(putBookmark, {
-    onMutate: async (postId) => {
-      await queryClient.cancelQueries({ queryKey: ['bookmarkPosts', lastId] });
-      const previousPosts = [...posts];
-
-      setPosts((prev) =>
-        prev.map((post) => {
-          if (post.postId === postId) {
-            return {
-              ...post,
-              isBookmarked: !post.isBookmarked,
-            };
-          }
-          return post;
-        }),
-      );
-
-      return { previousPosts } as MutationContext;
-    },
-    onError: (_, __, context) => {
-      if ((context as MutationContext)?.previousPosts) {
-        setPosts((context as MutationContext).previousPosts);
-      }
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['bookmarkPosts', lastId] });
-      queryClient.invalidateQueries({ queryKey: ['post', memberId] });
-    },
-  });
+  const { loveMutation, bookmarkMutation } = useLoveAndBookmark(posts, setPosts, memberId, lastId);
   return (
     <div className="flex flex-col gap-6">
       {isLoading
